+++ src/UIComponents/index.js
@@ -14,6 +14,7 @@ import ConfirmBox from "./atom/ConfirmBox";
 import Checkbox from "./atom/Checkbox";
 import TextArea from "./atom/TextArea";
 import StatusBar from "./atom/StatusBar";
+import Dropdown from "./atom/Dropdown";
 
 import EmptyState from "./molecules/EmptyState";
 import BlockerScreen from "./molecules/BlockerScreen";
@@ -37,6 +38,7 @@ export {
   EmptyState,
   CheckboxGroup,
   BlockerScreen,
+  Dropdown,
 };
 
 export default color;